requirements.txt
python-telegram-bot==20.7
aiohttp==3.8.4
gunicorn==20.1.0         # optional if you use WSGI; NOT needed for run_webhook